requirements.txt
# Core API
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3

# Quantum Computing
qiskit==0.45.0
qiskit-ibm-runtime==0.15.0
qiskit-aer
classiq==0.36.0
numpy==1.26.2
scipy==1.11.4

# Data Processing & GIS
pandas==2.1.3
geopandas==0.14.1
rasterio==1.3.9
shapely==2.0.2
scikit-learn==1.3.2

# Network & API
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0
websockets==12.0

# Database & Cache
sqlalchemy==2.0.23
redis[hiredis]==5.0.1

# Utilities
python-multipart==0.0.6
python-dateutil==2.8.2

# Performance (optional)
uvloop
httptools

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0